+++ src/messaging/messages.ts
@@ -53,7 +53,7 @@ type OptionsUpdated = {
 	type: 'nfe#optionsUpdated',
 }
 
-export type ToServiceWorkerMessage = RequestSiteDetails | OpenOptionsPage | CloseCurrentTab | RecordBlock | TrackUsageActivity | NotifyOptionsUpdated | SetSiteTheme | EnableSite | DisableSite | Snooze | ReadSnooze | SaveUsageLimits | ReadUsageLimits | PairFiniteSync | ReadFiniteSync | SyncFiniteNow | DisconnectFiniteSync;
+export type ToServiceWorkerMessage = RequestSiteDetails | OpenOptionsPage | CloseCurrentTab | TrackUsageActivity | NotifyOptionsUpdated | SetSiteTheme | EnableSite | DisableSite | Snooze | ReadSnooze | SaveUsageLimits | ReadUsageLimits | PairFiniteSync | ReadFiniteSync | SyncFiniteNow | DisconnectFiniteSync;
 
 // Request site details from service worker.
 type RequestSiteDetails = {
@@ -69,10 +69,6 @@ type OpenOptionsPage = {
 
 type CloseCurrentTab = {
 	type: 'closeCurrentTab',
-};
-
-type RecordBlock = {
-	type: 'recordBlock',
 };
 
 type TrackUsageActivity = {